Add functionality to operator buttons
When numbers are entered into the calculator now, as well as when
operator buttons are clicked, the result updates.

Two separate arrays were added to handle this, one to keep track of
the digits that are being pressed, and another that keeps track of each
operand as it is entered.

Depending on array length, the result bar is being updated differently.